placeholder for other shape nodes
script currently can deal with data of bezier and NURBS curves, but not for other shape types

added functionality only creates the related shape nodes, but not properly enumerate its data or physical description/attributes. i.e. it will create a nurbsSurface but not contain any CV data

this was added to make sure the shape nodes in the selection get properly processed for their corresponding transform DAG node